put address

+++ src/main/java/com/dambroski/foodDeliveryProject/User/UserServiceImpl.java
@@ -3,12 +3,14 @@ package com.dambroski.foodDeliveryProject.User;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.dambroski.foodDeliveryProject.Address.Address;
+import com.dambroski.foodDeliveryProject.error.UserNotFoundException;
 
 @Service
 public class UserServiceImpl implements UserService{
@@ -62,20 +64,31 @@ public class UserServiceImpl implements UserService{
 
 	@Override
 	public void deleteById(Long id) {
+		Optional<User> user = repository.findById(id);
+		if(user.isEmpty()) {
+			throw new UserNotFoundException("User not found");
+		}
 		repository.deleteById(id);
 		
 	}
 
 	@Override
 	public User editUser(User user,Long id) {
-		User newUser = repository.findById(id).get();
+		Optional<User> optionalUser = repository.findById(id);
+		if(optionalUser.isEmpty()) {
+			throw new UserNotFoundException("UserNotFound");
+		}
+		User newUser = optionalUser.get();
 		if(Objects.nonNull(user.getName()) && !"".equals(user.getName())) {
 			newUser.setName(user.getName());
 		}
 		if(Objects.nonNull(user.getEmail()) && !"".equals(user.getEmail())) {
 			newUser.setEmail(user.getEmail());
 		}
-		newUser.setBirthDate(user.getBirthDate());
+		if(Objects.nonNull(user.getBirthDate())) {
+			newUser.setBirthDate(user.getBirthDate());
+		}
+
 		
 		repository.save(newUser);
 		return newUser;
@@ -83,7 +96,8 @@ public class UserServiceImpl implements UserService{
 
 	@Override
 	public User addAddress(Address address, Long userId) {
-		User user = repository.findById(userId).get();
+		Optional<User> optionalUser = repository.findById(userId);
+		User user =  optionalUser.get();
 		if(user.getAddresses() == null) {
 			List<Address> listAddress = new ArrayList<>();
 			user.setAddresses(listAddress);
@@ -92,6 +106,31 @@ public class UserServiceImpl implements UserService{
 		address.setTypeId(user.getUserId());
 		user.getAddresses().add(address);
 		return repository.save(user);
+	}
+
+	@Override
+	public User editAddress(Address address,Long userId, Long addressId) {
+		Optional<User> optionalUser = repository.findById(userId);
+		if(optionalUser.isEmpty()) {
+			throw new UserNotFoundException("User not found");
+		}
+		User user = optionalUser.get();
+		List<Address> addresses = user.getAddresses();
+		for (Address addressList : addresses) {
+			if(addressList.getId() == addressId) {
+				if(Objects.nonNull(address.getAddress())) {
+					addressList.setAddress(address.getAddress());
+				}if(Objects.nonNull(address.getCity())) {
+					addressList.setCity(address.getCity());
+				}if(Objects.nonNull(address.getState())) {
+					addressList.setState(address.getState());
+				}
+	
+			}
+		}
+		user.setAddresses(addresses);
+		repository.save(user);
+		return user;
 	}
